Resolve duplicate save_results method and move load_model method
