+++ src/gnn_reco/components/utils.py
@@ -11,7 +11,7 @@ from torch_geometric.data.batch import Batch
 from sklearn.model_selection import train_test_split
 
 from gnn_reco.data.sqlite_dataset import SQLiteDataset
-from gnn_reco.models import Model
+
 
 # @TODO >>> RESOLVE DUPLICATION WRT. src/gnn_reco/models/training/{callbacks,trainers,utils}.py
 def make_train_validation_dataloader(db, selection, pulsemap, batch_size, FEATURES, TRUTH, num_workers, database_indices = None, persistent_workers = True, seed = 42):
@@ -41,28 +41,7 @@ def make_dataloader(db, selection, pulsemap, batch_size, FEATURES, TRUTH, num_wo
     dataloader = torch.utils.data.DataLoader(dataset, batch_size=batch_size, shuffle=True, num_workers=num_workers, 
                                             collate_fn=Batch.from_data_list,persistent_workers=persistent_workers,prefetch_factor=2)
     return dataloader
-def save_results(db, tag, results, archive,model):
-    db_name = db.split('/')[-1].split('.')[0]
-    path = archive + '/' + db_name + '/' + tag
-    os.makedirs(path, exist_ok = True)
-    results.to_csv(path + '/results.csv')
-    torch.save(model.cpu().state_dict(), path + '/' + tag + '.pkl')
-    print('Results saved at: \n %s'%path)
-    return
 # @TODO <<< RESOLVE DUPLICATION WRT. src/gnn_reco/models/training/{callbacks,trainers,utils}.py
-
-def load_model(db, tag, archive, detector, gnn, task, device):
-    db_name = db.split('/')[-1].split('.')[0]
-    path = archive + '/' + db_name + '/' + tag
-    model = Model(
-        detector=detector,
-        gnn=gnn,
-        tasks=[task],
-        device=device,
-    )
-    model.load_state_dict(torch.load(path + '/' + tag + '.pkl'))
-    model.eval()
-    return model
 
 def check_db_size(db):
     max_size = 5000000
